Change componets to be antd design and redesign edit/create profile pages
