+++ client/src/App.css
@@ -76,12 +76,20 @@ small {
   margin-bottom: 1.5rem;
 }
 
+.mb-3 {
+  margin-bottom: 1rem;
+}
+
 .mb-21 {
   margin-bottom: 0.5rem;
 }
 
 .mb-11 {
   margin-bottom: 0.25rem;
+}
+
+.ml-2 {
+  margin-left: 0.5rem;
 }
 
 .pb-31 {
