+++ client/src/App.js
@@ -16,10 +16,10 @@ import Landing from "./components/layout/landing";
 import Register from "./components/auth/register";
 import Login from "./components/auth/login";
 import Dashboard from "./components/dashboard";
-import CreateProfile from "./components/create-profile/CreateProfile";
+import CreateProfile from "./components/create-profile";
 import EditProfile from "./components/edit-profile/EditProfile";
-import AddExperience from "./components/add-credentials/AddExperience";
-import AddEducation from "./components/add-credentials/AddEducation";
+import AddExperience from "./components/add-credentials/add-experience";
+import AddEducation from "./components/add-credentials/add-education";
 import Profiles from "./components/profiles/Profiles";
 import Profile from "./components/profile/Profile";
 import NotFound from "./components/not-found/NotFound";
