+++ client/src/components/add-credentials/add-experience/index.js
@@ -1,22 +1,22 @@
 import React, { Component } from "react";
 import { Link, withRouter } from "react-router-dom";
-import TextFieldGroup from "../common/TextFieldGroup";
-import DateFieldGroup from "../common/DateFieldGroup";
-import TextAreaFieldGroup from "../common/TextAreaFieldGroup";
+import TextFieldGroup from "../../common/text-field-group";
+import DateFieldGroup from "../../common/date-field-group";
+import TextAreaFieldGroup from "../../common/text-area-field-group";
 import { Col, Row, Form, Button, Checkbox } from "antd";
 import { connect } from "react-redux";
 import PropTypes from "prop-types";
-import { addEducation } from "../../actions/profileActions";
+import { addExperience } from "../../../actions/profileActions";
 
 const FormItem = Form.Item;
 
-class AddEducation extends Component {
+class AddExperience extends Component {
   constructor(props) {
     super(props);
     this.state = {
-      school: "",
-      degree: "",
-      fieldofstudy: "",
+      company: "",
+      title: "",
+      location: "",
       from: "",
       to: "",
       current: false,
@@ -35,17 +35,17 @@ class AddEducation extends Component {
   onSubmit = e => {
     e.preventDefault();
 
-    const eduData = {
-      school: this.state.school,
-      degree: this.state.degree,
-      fieldofstudy: this.state.fieldofstudy,
+    const expData = {
+      company: this.state.company,
+      title: this.state.title,
+      location: this.state.location,
       from: this.state.from,
       to: this.state.to,
       current: this.state.current,
       description: this.state.description
     };
 
-    this.props.addEducation(eduData, this.props.history);
+    this.props.addExperience(expData, this.props.history);
   };
 
   onChange = e => {
@@ -63,7 +63,7 @@ class AddEducation extends Component {
     const { errors } = this.state;
 
     return (
-      <div className="add-education">
+      <div className="add-experience">
         <Row>
           <Col span="4">
             <Button>
@@ -73,32 +73,32 @@ class AddEducation extends Component {
         </Row>
         <Row type="flex" justify="center">
           <Col span={24}>
-            <h1 className="display-41 text-center1">Add Education</h1>
+            <h1 className="display-41 text-center1">Add Experience</h1>
             <p className="lead1 text-center1">
-              Add any school, bootcamp, etc that you have attended
+              Add any job or position that you have had in the past or current
             </p>
             <small className="display-block pb-31">* = required fields</small>
             <Form noValidate onSubmit={this.onSubmit}>
               <TextFieldGroup
-                placeholder="* School"
-                name="school"
-                value={this.state.school}
+                placeholder="* Company"
+                name="company"
+                value={this.state.company}
                 onChange={this.onChange}
-                error={errors.school}
+                error={errors.company}
               />
               <TextFieldGroup
-                placeholder="* Degree or Certification"
-                name="degree"
-                value={this.state.degree}
+                placeholder="* Job Title"
+                name="title"
+                value={this.state.title}
                 onChange={this.onChange}
-                error={errors.degree}
+                error={errors.title}
               />
               <TextFieldGroup
-                placeholder="* Field of Study"
-                name="fieldofstudy"
-                value={this.state.fieldofstudy}
+                placeholder="Location"
+                name="location"
+                value={this.state.location}
                 onChange={this.onChange}
-                error={errors.fieldofstudy}
+                error={errors.location}
               />
               <h6>From Date</h6>
               <DateFieldGroup
@@ -130,12 +130,12 @@ class AddEducation extends Component {
               </FormItem>
               <TextAreaFieldGroup
                 className="fs-1"
-                placeholder="Program Description"
+                placeholder="Job Description"
                 name="description"
                 value={this.state.description}
                 onChange={this.onChange}
                 error={errors.description}
-                info="Tell us about the program that you were in"
+                info="Tell us about the the position"
               />
               <Button
                 size="large"
@@ -153,8 +153,8 @@ class AddEducation extends Component {
   }
 }
 
-AddEducation.propTypes = {
-  addEducation: PropTypes.func.isRequired,
+AddExperience.propTypes = {
+  addExperience: PropTypes.func.isRequired,
   profile: PropTypes.object.isRequired,
   errors: PropTypes.object.isRequired
 };
@@ -166,5 +166,5 @@ const mapStateToProps = state => ({
 
 export default connect(
   mapStateToProps,
-  { addEducation }
-)(withRouter(AddEducation));
+  { addExperience }
+)(withRouter(AddExperience));
