+++ client/src/components/create-profile/index.js
@@ -1,15 +1,14 @@
 import React, { Component } from "react";
 import { connect } from "react-redux";
-import { Link, withRouter } from "react-router-dom";
+import { withRouter, Link } from "react-router-dom";
 import PropTypes from "prop-types";
 import TextFieldGroup from "../common/text-field-group";
 import TextAreaFieldGroup from "../common/text-area-field-group";
 import SelectListGroup from "../common/select-list-group";
-import { createProfile, getCurrentProfile } from "../../actions/profileActions";
-import { Row, Col, Button, Form } from "antd";
-import isEmpty from "../../validation/is-empty";
+import { createProfile } from "../../actions/profileActions";
+import { Row, Col, Form, Button } from "antd";
 
-class EditProfile extends Component {
+class CreateProfile extends Component {
   constructor(props) {
     super(props);
     this.state = {
@@ -18,7 +17,7 @@ class EditProfile extends Component {
       company: "",
       website: "",
       location: "",
-      status: "",
+      status: undefined,
       skills: "",
       githubusername: "",
       bio: "",
@@ -31,62 +30,9 @@ class EditProfile extends Component {
     };
   }
 
-  componentDidMount() {
-    this.props.getCurrentProfile();
-  }
-
   componentWillReceiveProps(nextProps) {
     if (nextProps.errors) {
       this.setState({ errors: nextProps.errors });
-    }
-
-    if (nextProps.profile.profile) {
-      const profile = nextProps.profile.profile;
-
-      // Bring skills array back to CSV
-      const skillsCSV = profile.skills.join(",");
-
-      // If profile field doesnt exist, make empty string
-      profile.company = !isEmpty(profile.company) ? profile.company : "";
-      profile.website = !isEmpty(profile.website) ? profile.website : "";
-      profile.location = !isEmpty(profile.location) ? profile.location : "";
-      profile.githubusername = !isEmpty(profile.githubusername)
-        ? profile.githubusername
-        : "";
-      profile.bio = !isEmpty(profile.bio) ? profile.bio : "";
-      profile.social = !isEmpty(profile.social) ? profile.social : {};
-      profile.twitter = !isEmpty(profile.social.twitter)
-        ? profile.social.twitter
-        : "";
-      profile.facebook = !isEmpty(profile.social.facebook)
-        ? profile.social.facebook
-        : "";
-      profile.linkedin = !isEmpty(profile.social.linkedin)
-        ? profile.social.linkedin
-        : "";
-      profile.youtube = !isEmpty(profile.social.youtube)
-        ? profile.social.youtube
-        : "";
-      profile.instagram = !isEmpty(profile.social.instagram)
-        ? profile.social.instagram
-        : "";
-
-      // Set component fields state
-      this.setState({
-        handle: profile.handle,
-        company: profile.company,
-        website: profile.website,
-        location: profile.location,
-        status: profile.status,
-        skills: skillsCSV,
-        githubusername: profile.githubusername,
-        bio: profile.bio,
-        twitter: profile.twitter,
-        facebook: profile.facebook,
-        linkedin: profile.linkedin,
-        youtube: profile.youtube,
-        instagram: profile.instagram
-      });
     }
   }
 
@@ -120,6 +66,7 @@ class EditProfile extends Component {
     const { errors, displaySocialInputs } = this.state;
 
     let socialInputs;
+
     if (displaySocialInputs) {
       socialInputs = (
         <React.Fragment>
@@ -184,7 +131,7 @@ class EditProfile extends Component {
     ];
 
     return (
-      <div className="edit-profile">
+      <div className="create-profile">
         <Row>
           <Col span="4">
             <Button>
@@ -248,7 +195,7 @@ class EditProfile extends Component {
                 onChange={this.onChange}
                 error={errors.skills}
                 info="Please use comma separated values (eg.
-                  HTML,CSS,JavaScript,PHP"
+                    HTML,CSS,JavaScript,PHP"
               />
               <TextFieldGroup
                 placeholder="Github Username"
@@ -296,9 +243,7 @@ class EditProfile extends Component {
   }
 }
 
-EditProfile.propTypes = {
-  createProfile: PropTypes.func.isRequired,
-  getCurrentProfile: PropTypes.func.isRequired,
+CreateProfile.propTypes = {
   profile: PropTypes.object.isRequired,
   errors: PropTypes.object.isRequired
 };
@@ -310,5 +255,5 @@ const mapStateToProps = state => ({
 
 export default connect(
   mapStateToProps,
-  { createProfile, getCurrentProfile }
-)(withRouter(EditProfile));
+  { createProfile }
+)(withRouter(CreateProfile));
